+++ crypto/fipsmodule/bcm.c
@@ -127,6 +127,7 @@
 #include "kem/kem.c"
 #include "md4/md4.c"
 #include "md5/md5.c"
+#include "ml_dsa/ml_dsa.c"
 #include "ml_kem/ml_kem.c"
 #include "modes/cbc.c"
 #include "modes/cfb.c"
