+++ crypto/fipsmodule/pqdsa/pqdsa.c
@@ -6,7 +6,7 @@
 
 #include "../delocate.h"
 #include "../../evp_extra/internal.h"
-#include "../crypto/ml_dsa/ml_dsa.h"
+#include "../ml_dsa/ml_dsa.h"
 #include "internal.h"
 
 // ML-DSA OIDs as defined within:
