+++ sources.cmake
@@ -82,6 +82,24 @@ set(
   crypto/fipsmodule/ecdsa/ecdsa_verify_tests.txt
   crypto/fipsmodule/kdf/test/kbkdf_counter.txt
   crypto/fipsmodule/kdf/test/sskdf.txt
+  crypto/fipsmodule/ml_dsa/kat/MLDSA_44_ACVP_keyGen.txt
+  crypto/fipsmodule/ml_dsa/kat/MLDSA_44_ACVP_sigGen.txt
+  crypto/fipsmodule/ml_dsa/kat/MLDSA_44_ACVP_sigVer.txt
+  crypto/fipsmodule/ml_dsa/kat/MLDSA_65_ACVP_keyGen.txt
+  crypto/fipsmodule/ml_dsa/kat/MLDSA_65_ACVP_sigGen.txt
+  crypto/fipsmodule/ml_dsa/kat/MLDSA_65_ACVP_sigVer.txt
+  crypto/fipsmodule/ml_dsa/kat/MLDSA_87_ACVP_keyGen.txt
+  crypto/fipsmodule/ml_dsa/kat/MLDSA_87_ACVP_sigGen.txt
+  crypto/fipsmodule/ml_dsa/kat/MLDSA_87_ACVP_sigVer.txt
+  crypto/fipsmodule/ml_dsa/kat/MLDSA_EXTMU_44_ACVP_sigGen.txt
+  crypto/fipsmodule/ml_dsa/kat/MLDSA_EXTMU_44_ACVP_sigVer.txt
+  crypto/fipsmodule/ml_dsa/kat/MLDSA_EXTMU_65_ACVP_sigGen.txt
+  crypto/fipsmodule/ml_dsa/kat/MLDSA_EXTMU_65_ACVP_sigVer.txt
+  crypto/fipsmodule/ml_dsa/kat/MLDSA_EXTMU_87_ACVP_sigGen.txt
+  crypto/fipsmodule/ml_dsa/kat/MLDSA_EXTMU_87_ACVP_sigVer.txt
+  crypto/fipsmodule/ml_dsa/kat/MLDSA_44_hedged_pure.txt
+  crypto/fipsmodule/ml_dsa/kat/MLDSA_65_hedged_pure.txt
+  crypto/fipsmodule/ml_dsa/kat/MLDSA_87_hedged_pure.txt
   crypto/fipsmodule/ml_kem/kat/mlkem512.txt
   crypto/fipsmodule/ml_kem/kat/mlkem768.txt
   crypto/fipsmodule/ml_kem/kat/mlkem1024.txt
@@ -102,24 +120,6 @@ set(
   crypto/kyber/kat/kyber512r3.txt
   crypto/kyber/kat/kyber768r3.txt
   crypto/kyber/kat/kyber1024r3.txt
-  crypto/ml_dsa/kat/MLDSA_44_ACVP_keyGen.txt
-  crypto/ml_dsa/kat/MLDSA_44_ACVP_sigGen.txt
-  crypto/ml_dsa/kat/MLDSA_44_ACVP_sigVer.txt
-  crypto/ml_dsa/kat/MLDSA_65_ACVP_keyGen.txt
-  crypto/ml_dsa/kat/MLDSA_65_ACVP_sigGen.txt
-  crypto/ml_dsa/kat/MLDSA_65_ACVP_sigVer.txt
-  crypto/ml_dsa/kat/MLDSA_87_ACVP_keyGen.txt
-  crypto/ml_dsa/kat/MLDSA_87_ACVP_sigGen.txt
-  crypto/ml_dsa/kat/MLDSA_87_ACVP_sigVer.txt
-  crypto/ml_dsa/kat/MLDSA_EXTMU_44_ACVP_sigGen.txt
-  crypto/ml_dsa/kat/MLDSA_EXTMU_44_ACVP_sigVer.txt
-  crypto/ml_dsa/kat/MLDSA_EXTMU_65_ACVP_sigGen.txt
-  crypto/ml_dsa/kat/MLDSA_EXTMU_65_ACVP_sigVer.txt
-  crypto/ml_dsa/kat/MLDSA_EXTMU_87_ACVP_sigGen.txt
-  crypto/ml_dsa/kat/MLDSA_EXTMU_87_ACVP_sigVer.txt
-  crypto/ml_dsa/kat/MLDSA_44_hedged_pure.txt
-  crypto/ml_dsa/kat/MLDSA_65_hedged_pure.txt
-  crypto/ml_dsa/kat/MLDSA_87_hedged_pure.txt
   crypto/ocsp/test/aws/certs.txt
   crypto/ocsp/test/aws/certs_revoked.txt
   crypto/ocsp/test/aws/certs_unknown.txt
